+++ config/polyfills.js
@@ -1,2 +1,2 @@
 global.regeneratorRuntime = require("regenerator-runtime");
-global.fetch = require("jest-fetch-mock");
+require("jest-fetch-mock").enableMocks();