+++ src/KeyManager.ts
@@ -277,14 +277,16 @@ export class KeyManager {
    *                           computed as `sha1(private key + public key)`.
    * @param {string} params.password The password that will decrypt that secret
    * @param {string} params.authServer The URL of the authentication server
-   * @param {string} params.account The authenticating public key. If not
+   * @param {string} params.authServerKey Check the challenge transaction
+   *                                for this key as source and signature.
+   * @param {string} [params.account] The authenticating public key. If not
    *                                provided, then the signers's public key will
    *                                be used instead.
    * @returns {Promise<string>} authToken JWT
    */
   // tslint:enable max-line-length
   public async fetchAuthToken(params: GetAuthTokenParams): Promise<AuthToken> {
-    const { id, password, authServer } = params;
+    const { id, password, authServer, authServerKey } = params;
     let { account } = params;
 
     // throw errors for missing params
@@ -372,6 +374,30 @@ export class KeyManager {
           firstTransaction.sequence
         }`,
       );
+    }
+
+    if (authServerKey) {
+      if (firstTransaction.source !== authServerKey) {
+        throw new Error(
+          `Signing key doesn't match: Expected ${authServerKey} but got
+          ${firstTransaction.source}`,
+        );
+      }
+
+      if (
+        !firstTransaction.signatures.some((signature) =>
+          signature
+            .hint()
+            .equals(
+              StellarSdk.Keypair.fromPublicKey(authServerKey).signatureHint(),
+            ),
+        )
+      ) {
+        throw new Error(
+          `Signing key doesn't match: Expected ${authServerKey} but got
+          something different`,
+        );
+      }
     }
 
     const signedTransaction = await keyHandler.signTransaction({